Fix overflow when hovering navigation buttons
add a small margin bottom to the button part of the swiper.
This prevents an overflow, with the scroll-bar, on small screens
when hovering the navigation buttons.
This overflow occurs because the arrow rotates on hover.
Adding a margin to the pagination div prevents this overflow as it
increases the height of the navigation container.

Signed-off-by: Pascal Brand <pascal.brand38@gmail.com>